pips-agent/requirements.txt
claude-agent-sdk>=0.1.18
python-dotenv>=1.0.0
PyYAML>=6.0
opencv-python>=4.8.0
numpy>=1.24.0
pytesseract>=0.3.10
scikit-learn>=1.3.0
